+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Release date: ?
 
  - added Http\Middleware\JsonBodyParserMiddleware and Http\Exception\BodyParserException
  - allow empty argument in Collection\Structure\AbstractStructure::create()
+ - added Blueprint\Collection\Structure
 
 VERSION 4.0.0-RC1
 -----------