+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Release date: ?
  - added method getKeys() and static keys() to Collection\Structure\AbstractStructure
  - fixed edge case where singleton binding as string were not used properly in Di\Container and Di\Binding\Singleton
  - allow the usage of binding when autoWiring is on in Di\Container
+ - allow resolution of interfaces dependencies via bindings in Di\InterfaceResolver
 
 VERSION 4.0.0-RC1
 -----------