+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Release date: ?
  - added Di\Exception\InfiniteLoopResolutionException
  - moved Di\AbstractBinding and Di\BindingInterface under Di\Binding folder
  - renamed method getKeys() to getStructureKeys() in Collection\Structure\AbstractStructure to be consistent
-
+ - added method getHandlers() to Blueprint\Http\Stack and update Http\Stack and Http\Request\Route
 VERSION 4.0.0-RC1
 -----------
 Release date: 2019-03-30