+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Release date: ?
  - moved Di\AbstractBinding and Di\BindingInterface under Di\Binding folder
  - renamed method getKeys() to getStructureKeys() in Collection\Structure\AbstractStructure to be consistent
  - added method getHandlers() to Blueprint\Http\Stack and update Http\Stack and Http\Request\Route
+ - removed deprecated components in Common: Reflection, Pagination, ClassFinder, TextUtils and TimeExpression
+ 
 VERSION 4.0.0-RC1
 -----------
 Release date: 2019-03-30