+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 VERSION 4.0.0-RC2
 -----------
-Release date: ?
+Release date: 2019-05-24
 
  - added Http\Middleware\JsonBodyParserMiddleware and Http\Exception\BodyParserException
  - allow empty argument in Collection\Structure\AbstractStructure::create()